+++ src/main/java/um/tesoreria/core/hexagonal/guarani/guaraniBeneficio/infrastructure/web/dto/GuaraniBeneficioRequest.java
@@ -3,6 +3,7 @@ package um.tesoreria.core.hexagonal.guarani.guaraniBeneficio.infrastructure.web.
 import lombok.Data;
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
+import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.NotNull;
 import java.math.BigDecimal;
 
@@ -13,6 +14,7 @@ public class GuaraniBeneficioRequest {
 
     @NotNull
     @DecimalMin(value = "0", inclusive = true)
-    @DecimalMax(value = "100", inclusive = true)
+    @DecimalMax(value = "1", inclusive = true)
+    @Digits(integer = 1, fraction = 2)
     private BigDecimal porcentajeBeneficio;
 }